Skip dumping backtraces for cached processes
This change updates dumpstate to skip dumping backtraces for cached
processes. These processes are less likely to be causing
bugreport-necessitating issues, and if they are frozen, will cause
backtrace dumps to time out anyways.

This results in a ~50% decrease (21.327s -> 10.568s) in the time taken
to run DUMP TRACES.

Test: On a local device, check with bugreportz that there are no more
      timed-out backtrace dumps (i.e. that the "Dump failed, likely due
      to a timeout" failure no longer occurs).
Test: Run health/bugreport/capturebugreport on an internal build and
      verify that the DUMP TRACES section is much shorter.
Bug: 276451949
Google: 2528109
Change-Id: Iecb6c2168a0dccefecf3b92be24c647eaf729556